Fixed hoNDInterpolater and friends to correctly use const arrays. Changed T1 fit to use 5th order spline for image interpolation

+++ toolboxes/core/cpu/hoNDArray.h
@@ -106,7 +106,7 @@ namespace Gadgetron{
   public:
 
     typedef NDArray<T> BaseClass;
-    typedef float coord_type;
+    typedef double coord_type;
     typedef T value_type;
     using iterator = T*;
     using const_iterator = const T*;
